# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentEdgeDefinition": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target"
        ],
        "title": "AgentEdgeDefinition",
        "type": "object"
      },
      "AgentGraphDefinition": {
        "additionalProperties": false,
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/AgentEdgeDefinition"
            },
            "maxItems": 200,
            "title": "Edges",
            "type": "array"
          },
          "entry_point": {
            "title": "Entry Point",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/AgentNodeDefinition"
            },
            "maxItems": 80,
            "minItems": 1,
            "title": "Nodes",
            "type": "array"
          },
          "output_node": {
            "title": "Output Node",
            "type": "string"
          }
        },
        "required": [
          "entry_point",
          "output_node",
          "nodes",
          "edges"
        ],
        "title": "AgentGraphDefinition",
        "type": "object"
      },
      "AgentNodeDefinition": {
        "additionalProperties": false,
        "properties": {
          "handler": {
            "pattern": "^[a-z][a-z0-9_]{1,80}$",
            "title": "Handler",
            "type": "string"
          },
          "id": {
            "pattern": "^[a-z][a-z0-9_]{1,80}$",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "deterministic",
              "llm"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "prompt_key": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9-]{1,119}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Key"
          },
          "resource_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeResourcePolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "response_schema": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Schema"
          },
          "runtime_input_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeInputPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "snapshot_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeSnapshotPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "task": {
            "anyOf": [
              {
                "maxLength": 240,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          }
        },
        "required": [
          "id",
          "label",
          "kind",
          "handler"
        ],
        "title": "AgentNodeDefinition",
        "type": "object"
      },
...
      "AgentVersionDetail": {
        "properties": {
          "agent_system_id": {
            "title": "Agent System Id",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "definition": {
            "additionalProperties": true,
            "title": "Definition",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "agent_system_id",
          "version",
          "definition",
          "content_hash",
          "created_at"
        ],
        "title": "AgentVersionDetail",
        "type": "object"
      },
...
      "AutoEvalManifest": {
        "additionalProperties": false,
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/AgentGraphDefinition"
          },
          "prompt": {
            "$ref": "#/components/schemas/ManifestPrompt"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "system": {
            "$ref": "#/components/schemas/ManifestSystem"
          }
        },
        "required": [
          "schema_version",
          "system",
          "graph",
          "prompt"
        ],
        "title": "AutoEvalManifest",
        "type": "object"
      },
...
      "CreateAgentVersionRequest": {
        "additionalProperties": false,
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/AgentGraphDefinition"
          }
        },
        "required": [
          "definition"
        ],
        "title": "CreateAgentVersionRequest",
        "type": "object"
      },
...
      "CreatePromptVersionRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "CreatePromptVersionRequest",
        "type": "object"
      },
...
      "EffectiveHandlersResponse": {
        "additionalProperties": false,
        "properties": {
          "deterministic": {
            "items": {
              "type": "string"
            },
            "title": "Deterministic",
            "type": "array"
          },
          "llm": {
            "items": {
              "type": "string"
            },
            "title": "Llm",
            "type": "array"
          },
          "system_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Key"
          }
        },
        "required": [
          "system_key",
          "deterministic",
          "llm"
        ],
        "title": "EffectiveHandlersResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportSource": {
        "additionalProperties": false,
        "properties": {
          "display_path": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Path"
          },
          "kind": {
            "enum": [
              "local",
              "github"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "ImportSource",
        "type": "object"
      },
      "ManifestPrompt": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "ManifestPrompt",
        "type": "object"
      },
      "ManifestSystem": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "maxLength": 10000,
            "title": "Description",
            "type": "string"
          },
          "input_template": {
            "additionalProperties": true,
            "title": "Input Template",
            "type": "object"
          },
          "key": {
            "pattern": "^[a-z][a-z0-9-]{1,119}$",
            "title": "Key",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name"
        ],
        "title": "ManifestSystem",
        "type": "object"
      },
...
      "NodeResourcePolicy": {
        "additionalProperties": false,
        "description": "Server-owned producer contract for a consumer node's persisted resource.",
        "properties": {
          "evaluation_mode": {
            "const": "locked",
            "default": "locked",
            "title": "Evaluation Mode",
            "type": "string"
          },
          "producer_node_id": {
            "pattern": "^[a-z][a-z0-9_]{1,80}$",
            "title": "Producer Node Id",
            "type": "string"
          },
          "producer_output_key": {
            "pattern": "^[a-z][a-z0-9_]{1,80}$",
            "title": "Producer Output Key",
            "type": "string"
          },
          "producer_snapshot_kind": {
            "enum": [
              "state",
              "external_observation",
              "node_output"
            ],
            "title": "Producer Snapshot Kind",
            "type": "string"
          },
          "producer_system_key": {
            "pattern": "^[a-z][a-z0-9-]{1,119}$",
            "title": "Producer System Key",
            "type": "string"
          },
          "product_key": {
            "pattern": "^[a-z][a-z0-9-]{1,119}$",
            "title": "Product Key",
            "type": "string"
          },
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          },
          "resource_key": {
            "pattern": "^[a-z][a-z0-9_]{1,80}$",
            "title": "Resource Key",
            "type": "string"
          },
          "runtime_mode": {
            "default": "current",
            "enum": [
              "current",
              "locked"
            ],
            "title": "Runtime Mode",
            "type": "string"
          },
          "schema_version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "product_key",
          "resource_key",
          "producer_system_key",
          "producer_node_id",
          "producer_output_key",
          "producer_snapshot_kind"
        ],
        "title": "NodeResourcePolicy",
        "type": "object"
      },
...
      "PromptVersionDetail": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "prompt_id": {
            "title": "Prompt Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "prompt_id",
          "version",
          "content",
          "content_hash",
          "created_at"
        ],
        "title": "PromptVersionDetail",
        "type": "object"
      },
...
      "SystemImportCommitRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Expected Digest",
            "type": "string"
          },
          "expected_graph_version": {
            "minimum": 1.0,
            "title": "Expected Graph Version",
            "type": "integer"
          },
          "expected_prompt_version": {
            "minimum": 1.0,
            "title": "Expected Prompt Version",
            "type": "integer"
          },
          "expected_system_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected System Id"
          },
          "expected_system_key": {
            "pattern": "^[a-z][a-z0-9-]{1,119}$",
            "title": "Expected System Key",
            "type": "string"
          },
          "manifest": {
            "$ref": "#/components/schemas/AutoEvalManifest"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImportSource"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "manifest",
          "expected_digest",
          "expected_system_key",
          "expected_graph_version",
          "expected_prompt_version"
        ],
        "title": "SystemImportCommitRequest",
        "type": "object"
      },
      "SystemImportCompatibility": {
        "additionalProperties": false,
        "properties": {
          "evaluation_scoring": {
            "enum": [
              "exact_json",
              "registered"
            ],
            "title": "Evaluation Scoring",
            "type": "string"
          },
          "handlers": {
            "items": {
              "type": "string"
            },
            "title": "Handlers",
            "type": "array"
          },
          "runnable": {
            "title": "Runnable",
            "type": "boolean"
          }
        },
        "required": [
          "runnable",
          "evaluation_scoring",
          "handlers"
        ],
        "title": "SystemImportCompatibility",
        "type": "object"
      },
      "SystemImportCounts": {
        "additionalProperties": false,
        "properties": {
          "deterministic_nodes": {
            "title": "Deterministic Nodes",
            "type": "integer"
          },
          "edges": {
            "title": "Edges",
            "type": "integer"
          },
          "llm_nodes": {
            "title": "Llm Nodes",
            "type": "integer"
          },
          "nodes": {
            "title": "Nodes",
            "type": "integer"
          }
        },
        "required": [
          "nodes",
          "edges",
          "llm_nodes",
          "deterministic_nodes"
        ],
        "title": "SystemImportCounts",
        "type": "object"
      },
      "SystemImportInspectRequest": {
        "additionalProperties": false,
        "properties": {
          "github_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Url"
          },
          "manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AutoEvalManifest"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImportSource"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "SystemImportInspectRequest",
        "type": "object"
      },
      "SystemImportPreview": {
        "additionalProperties": false,
        "properties": {
          "compatibility": {
            "$ref": "#/components/schemas/SystemImportCompatibility"
          },
          "counts": {
            "$ref": "#/components/schemas/SystemImportCounts"
          },
          "creates_system": {
            "title": "Creates System",
            "type": "boolean"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "existing_system_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing System Id"
          },
          "graph_version_created": {
            "title": "Graph Version Created",
            "type": "boolean"
          },
          "manifest": {
            "$ref": "#/components/schemas/AutoEvalManifest"
          },
          "next_graph_version": {
            "title": "Next Graph Version",
            "type": "integer"
          },
          "next_prompt_version": {
            "title": "Next Prompt Version",
            "type": "integer"
          },
          "prompt_version_created": {
            "title": "Prompt Version Created",
            "type": "boolean"
          },
          "source": {
            "$ref": "#/components/schemas/ImportSource"
          },
          "target_graph_version": {
            "title": "Target Graph Version",
            "type": "integer"
          },
          "target_prompt_version": {
            "title": "Target Prompt Version",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "digest",
          "manifest",
          "source",
          "warnings",
          "existing_system_id",
          "next_graph_version",
          "next_prompt_version",
          "target_graph_version",
          "target_prompt_version",
          "graph_version_created",
          "prompt_version_created",
          "creates_system",
          "counts",
          "compatibility"
        ],
        "title": "SystemImportPreview",
        "type": "object"
      },
      "SystemImportResult": {
        "additionalProperties": false,
        "properties": {
          "agent_system_id": {
            "title": "Agent System Id",
            "type": "string"
          },
          "created_system": {
            "title": "Created System",
            "type": "boolean"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "dataset_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Version Id"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "graph_version": {
            "title": "Graph Version",
            "type": "integer"
          },
          "graph_version_created": {
            "title": "Graph Version Created",
            "type": "boolean"
          },
          "graph_version_id": {
            "title": "Graph Version Id",
            "type": "string"
          },
          "prompt_id": {
            "title": "Prompt Id",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "integer"
          },
          "prompt_version_created": {
            "title": "Prompt Version Created",
            "type": "boolean"
          },
          "prompt_version_id": {
            "title": "Prompt Version Id",
            "type": "string"
          }
        },
        "required": [
          "agent_system_id",
          "graph_version_id",
          "graph_version",
          "graph_version_created",
          "prompt_id",
          "prompt_version_id",
          "prompt_version",
          "prompt_version_created",
          "dataset_id",
          "dataset_version_id",
          "created_system",
          "digest"
        ],
        "title": "SystemImportResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AutoEval API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent-system-versions/{version_id}": {
      "get": {
        "operationId": "agent_version_detail_api_agent_system_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentVersionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent Version Detail"
      }
    },
...
    "/api/agent-systems/{agent_system_id}/versions": {
      "post": {
        "operationId": "add_agent_version_api_agent_systems__agent_system_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_system_id",
            "required": true,
            "schema": {
              "title": "Agent System Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentVersionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Agent Version"
      }
    },
...
    "/api/prompt-versions/{version_id}": {
      "get": {
        "operationId": "prompt_version_detail_api_prompt_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptVersionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prompt Version Detail"
      }
    },
    "/api/prompts/{prompt_id}/versions": {
      "post": {
        "operationId": "add_prompt_version_api_prompts__prompt_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "prompt_id",
            "required": true,
            "schema": {
              "title": "Prompt Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptVersionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Prompt Version"
      }
    },
...
    "/api/system-imports/commit": {
      "post": {
        "operationId": "commit_system_import_api_system_imports_commit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemImportCommitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit System Import"
      }
    },
    "/api/system-imports/handlers": {
      "get": {
        "operationId": "effective_handlers_api_system_imports_handlers_get",
        "parameters": [
          {
            "in": "query",
            "name": "system_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^[a-z][a-z0-9-]{1,119}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "System Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveHandlersResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Effective Handlers"
      }
    },
    "/api/system-imports/inspect": {
      "post": {
        "operationId": "inspect_system_import_api_system_imports_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemImportInspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemImportPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect System Import"
      }
    },
...
  }
}